requirements.txt
kafka-python
pyspark
avro-python3
